requirements.txt
asgiref==3.2.7
Django==2.2.5
django-cors-headers==2.2.0
django-discover-runner==1.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==19.9.0
Pillow==7.0.0
PyJWT==1.7.1
pytz==2019.2
six==1.14.0
sqlparse==0.3.0
